refactor: add usage to scripts
